% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateTailBeats.R
\name{.processBatch}
\alias{.processBatch}
\title{Helper function to process a single batch}
\usage{
.processBatch(
  batch,
  valid_data,
  batch_size,
  buffer_samples,
  n_rows_valid,
  sampling_freq,
  min.freq.Hz,
  max.freq.Hz,
  ridge.only,
  power.ratio.threshold
)
}
\description{
Helper function to process a single batch
}
\note{
This function is intended for internal use within the \code{nautilus} package.
}
\keyword{internal}
