% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_files_with_functions}
\alias{find_files_with_functions}
\title{Get names of files that will be contain function names}
\usage{
find_files_with_functions(path = here::here())
}
\arguments{
\item{path}{Presumes that the current working directory is the main folder of the current R project/package.}
}
\value{
Character string of file paths
}
\description{
Get names of files that will be contain function names
}
