% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracks_per_target.R
\name{get_tracks_per_target}
\alias{get_tracks_per_target}
\title{Get number of tracks assigned to each target over time}
\usage{
get_tracks_per_target(assignmentData, truthData = NULL)
}
\arguments{
\item{assignmentData}{the output of target_assignment()}

\item{truthData}{the input truth data (default=NULL). If you don't pass this, the return is at sensor times. If you DO pass this, return is at truth data times.}
}
\value{
dataframe containing lat, lon, alt, heading, time, truthID, and numTracksOnTarget
}
\description{
Takes assignmentData, and figures out how many tracks were assigned to each target over time. It does this by looking for when track segments start and stop. NOTE! false tracks will NOT interrupt a track, so if a target is tracked at time t=1 and t=10 (with false tracks in between), this will assume it was tracked the whole time
}
\keyword{internal}
