% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nautilus-package.R
\docType{package}
\name{nautilus-package}
\alias{nautilus}
\alias{nautilus-package}
\title{nautilus: Navigate Function Development}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
