% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nautilus.R
\name{nautilus}
\alias{nautilus}
\title{Nautilus Aka Snail Plot}
\usage{
nautilus(x, col, lwd = 1, FUN = c("qlog10", "asinh.co", "asinh",
  "log10", "lin"), cofactor = 0.2, seed = 25, p = 1, coef = 0,
  untie = 0, switch.cols = FALSE, distfun = dist,
  hclustfun = hclust, NAUT, lwd.shell = 0.45, col.shell = "black",
  cex = 0.45, adj = 0.2, plot.major = TRUE, lwd.major.l = 0.15,
  lwd.major.g = 0.45, col.major = "gray", plot.minor = TRUE,
  lwd.minor.l = 0.05, lwd.minor.g = 0.15, col.minor = "gray80",
  plot.axis = TRUE, cex.axis = 0.35, padj.axis = -4.5,
  lwd.major.axis = 0.75, lwd.major.ticks = 0.75, tck.major = -0.015,
  lwd.minor.axis = 0.05, lwd.minor.ticks = 0.5, tck.minor = -0.0075,
  plot.vertical = TRUE, lwd.vertical = 0.15, col.vertical = "gray")
}
\arguments{
\item{x}{numeric matrix of the values to be plotted with observations in rows and variables in columns}

\item{col}{line colors to be used, e.g. according to a grouping factor, vector of type character and same length as rows in x}

\item{lwd}{line width, Default: 1}

\item{FUN}{function to be used for log or log-like transformation, if "lin" no transformation is performed, Default: c("qlog10", "asinh.co", "asinh", "log10", "lin")}

\item{cofactor}{for asinh.co as e.g. used for (non-negative) cytometry data transformation, Default: 0.2}

\item{seed}{for randomizing order of plotted lines (to prevent masking), may be turned of by setting to NA, Default: 25}

\item{p}{quantile of the data which is taken to order variables, e.g. p = 0.5 for medians or p = 1 for maxima, Default: 1}

\item{coef}{used to remove outliers by IQR rule when set to positive value, e.g. 1.5 as in \code{\link[grDevices]{boxplot.stats}}, Default: 0}

\item{untie}{this provides an experimental option to reorder fractions of the plot in order to bring correlated variables closer. It uses hierarchical clustering for reordering of at least 3 variables per fraction., Default: 0}

\item{switch.cols}{logical whether to reverse reordering within each fraction, Default: FALSE}

\item{distfun}{distance function used in clustering for within fraction reordering, Default: \code{\link[stats]{dist}}}

\item{hclustfun}{linkage function used in clustering for within fraction reordering, Default: \code{\link[stats]{hclust}}}

\item{NAUT}{an object as returned from a previous call to the function. This will override calculation of all dimensions of the new plot, including variable ordering. This can come handy if data shall be plotted at same scale but different figures}

\item{lwd.shell}{width of the outer line, Default: 0.45}

\item{col.shell}{color of the outer line, Default: 'black'}

\item{cex}{character expansion factor for the labels, Default: 0.45}

\item{adj}{label adjustment, see \code{\link[graphics]{par}}, Default: 0.2}

\item{plot.major}{logical whether to plot main scale lines, Default: TRUE}

\item{lwd.major.l}{width of the smallest main scale line, Default: 0.15}

\item{lwd.major.g}{width of the greatest main scale line. Line widths in-between will be ramped according to these two values, Default: 0.45}

\item{col.major}{color of main scale lines, Default: 'gray'}

\item{plot.minor}{logical whether to plot minor scale lines, only effective for log10- or asinh-transformed data, Default: TRUE}

\item{lwd.minor.l}{width of the smallest minor scale line, Default: 0.05}

\item{lwd.minor.g}{width of the greatest minor scale line. Line widths in-between will be ramped according to these two values, Default: 0.15}

\item{col.minor}{color of minor scale lines, Default: 'gray80'}

\item{plot.axis}{logical whether to plot axis, Default: TRUE}

\item{cex.axis}{expansion factor for axis labels, Default: 0.35}

\item{padj.axis}{adjustment for each tick label, see \code{\link[graphics]{axis}}, Default: -4.5}

\item{lwd.major.axis}{line width for the axis line, Default: 0.75}

\item{lwd.major.ticks}{line widths the main tick marks, Default: 0.75}

\item{tck.major}{length of main tick marks, Default: -0.015}

\item{lwd.minor.axis}{line width for the axis line, Default: 0.05}

\item{lwd.minor.ticks}{line widths for the minor tick marks, Default: 0.5}

\item{tck.minor}{length of minor tick marks, Default: -0.0075}

\item{plot.vertical}{logical whether to plot vertical lines, Default: TRUE}

\item{lwd.vertical}{width of vertical lines, Default: 0.15}

\item{col.vertical}{color of vertical lines, Default: 'gray'}
}
\value{
Invisibly, a list object with following fields:
\describe{
  \item{\code{FUN}}{transformation used}
  \item{\code{SCALES}}{matrix containing plotting coordinates for the shell}
  \item{\code{ORD}}{used order of variables}
  \item{\code{SHIFT}}{constant to move all values on positive scale}
  \item{\code{main.ticks}}{main tick mark coordinates}
  \item{\code{minor.ticks}}{minor tick mark coordinates}
  \item{\code{log.ticks}}{main tick labels on log scale}
}
}
\description{
Draws a modified stars (or radar) plot.
}
\details{
Starting with 10 to 15 variables, the stars (or radar) plot can be made to resemble a nautilus (\eqn{Nautilus belauensis}). With more than 50 markers it will however be difficult to grasp the actual expression profiles. Besides the standard log10 transformation, the plotting function offers asinh-based transformations in order to bring the data properly on scale.
}
\examples{
\dontrun{

# mock up data with e.g. 100 samples total, four groups and 30 variables
set.seed(25)
g <- rep(1:4, each = 25)
sc <- 10^sample(seq(0, 3, length.out = 100), 30, TRUE)
y <- sapply(sc, function(x) rnorm(100, x, x / 2) + x * rep(sample.int(4), each = 25))
colnames(y) <- colnames(y, do.NULL = FALSE, prefix = "VAR")
col <- palette()[g + 1] # according to group code

# asinh.co allows display of negative data on "quasi-" logarithmic scale
nautilus(y, col = col, FUN = "asinh.co", padj.axis = -5.5)
# pair correlated variables
nautilus(y, col = col, FUN = "asinh.co", cofactor = 10, padj.axis = -5.5, untie = 4)
nautilus(y, col = col, FUN = "asinh.co", cofactor = 10, padj.axis = -5.5, untie = 4, 
            switch.cols = TRUE, plot.major = FALSE, plot.minor = FALSE, plot.vertical = FALSE)
# default asinh
nautilus(y, col = col, FUN = "asinh", padj.axis = -5.5)
# quasi-log10 
nautilus(y, col = col, FUN = "qlog10", padj.axis = -5.5)

# default log10
nautilus(y, col = col, FUN = "log10")

# log-normal data
set.seed(25)
g <- rep(1:4, each = 25)
sc <- 10^sample(seq(-2, 3, length.out = 100), 30, TRUE)
y <- sapply(sc, function(x) rnorm(100, x, abs(x))  +  x * rep(sample.int(4), each = 25))
colnames(y) <- colnames(y, do.NULL = FALSE, prefix = "VAR")
col <- palette()[g + 1]
nautilus(y, col = col, FUN = "log10")

# linear scale
nautilus(y, col = col, FUN = "lin", adj = 500)

# normal data
set.seed(25)
g <- rep(1:4, each = 25)
sc <- sample(seq(-1, 5, length.out = 100), 30, TRUE)
y <- 100 + sapply(sc, function(x) rnorm(100, x, abs(x) / 2)  +  x * rep(sample.int(4), each = 25))
colnames(y) <- colnames(y, do.NULL = FALSE, prefix = "VAR")
col <- palette()[g + 1]

nautilus(y, col = col, FUN = "lin", p = 0.5, padj.axis = -5.5)

}
}
\seealso{
\code{\link[graphics]{stars}}, \code{\link[graphics]{axis}}
}
