% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nautilus_interface.R
\name{nautilus_interface}
\alias{nautilus_interface}
\title{Nautilus interface function (shiny)}
\usage{
nautilus_interface()
}
\value{
none
}
\description{
Function to create shiny interface
}
\keyword{internal}
