% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_track_status.R
\name{plot_compare_track_status}
\alias{plot_compare_track_status}
\title{Compares the track status from different assignmentDatas}
\usage{
plot_compare_track_status(
  scenario1,
  scenario2,
  nameList = NA,
  showFalseTracks = FALSE
)
}
\arguments{
\item{scenario1}{must contain track-to-truth assignments (the output of target_assignment()) and truthData}

\item{scenario2}{must contain track-to-truth assignments (the output of target_assignment()) and truthData}

\item{nameList}{title of each scenario in a list, (default = the scenario names)}

\item{showFalseTracks}{will plot false tracks for comparison (default=FALSE)}
}
\value{
ggplot object
}
\description{
This figure compares the target assignments from two methods.  Each subfigure is similar to plot_track_status().  Areas of disagreement between the two methods are shown in bold.  If there are differences in points flagged as false tracks they are shown as triangles.
}
\examples{
\dontrun{
pointScenario=scenarioMaker::example2_scenario \%>\%
 target_assignment("point",cutoff=100)

gaussScenario=scenarioMaker::example2_scenario \%>\%
 target_assignment("windowGauss",cutoff=100, window=10)

plot_compare_track_status(scenario1 = pointScenario,
  scenario2 = gaussScenario,
  nameList = list("Point Method", "Gauss Window Method"),
  showFalseTracks = FALSE)
  }
}
