% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polar_error.R
\name{plot_polar_error}
\alias{plot_polar_error}
\title{Plot [location error / bearing error / range error] by [target aspect / bearing to target]}
\usage{
plot_polar_error(scenario, angleTerm = "bearing", rTerm = "locationError",
  colorTerm = "tgtAssigned", doFacet = FALSE, plotFalseTracks = FALSE)
}
\arguments{
\item{scenario}{MUST contain assignmentData (from target_assignment())}

\item{angleTerm}{(string) the variable plotted as an angle. Currently there are two options:
\itemize{
 \item{"bearing" - the bearing to the target (are you looking at something straight ahead or to your side?)}
 \item{"aspect" - the target's aspect (is it facing you?)}}}

\item{rTerm}{(string) the variable plotted radially. Currently there are three options:
\itemize{
 \item{"locationError"}
 \item{"bearingError"}
 \item{"downrangeError"}}}

\item{colorTerm}{(string) the variable to be split by color. Currently there are two options:
\itemize{
 \item{"tgtAssigned"}
 \item{"trackNum"}}}

\item{doFacet}{(default FALSE) boolean.  Determines whether the graph will also facet by colorTerm}

\item{plotFalseTracks}{(default FALSE) boolead. Determines whether graph will show fast track points or hide them}
}
\value{
ggplot object
}
\description{
This figure plots one of several choices for system error (location error, range error, bearing error) against the target aspect or the bearing to the target
}
\examples{
plot_polar_error(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100),angleTerm="bearing", rTerm = 'locationError',doFacet = TRUE, plotFalseTracks=FALSE)
}
