% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatterplot_with_density.R
\name{plot_scatterplot_with_density}
\alias{plot_scatterplot_with_density}
\title{Scatterplot with density sidebars}
\usage{
plot_scatterplot_with_density(scenario, xValue, yValue)
}
\arguments{
\item{scenario}{MUST contain assignmentData (from target_assignment())}

\item{xValue}{the variable to be plotted on the x-axis. Can be any of the columns in assignmentData. Must be put into a string. We generally recommend "bearingError", however any continuous value can be used.}

\item{yValue}{the variable to be plotted on the y-axis. Can be any of the columns in assignmentData. Must be put into a string. We generally recommend "downrangeError", however any continuous value can be used.}
}
\value{
layot of ggplot objects
}
\description{
This figure plots any two continuous variables together to check for systematic bias.  We recommend using bearingError and downrangeError.

  Note: This is a more generic function and therefore can be used for other comparisons. However, this means that the user must specify that the data to be plotted should not include points labeled as false tracks.
}
\examples{
myScenario=scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100)
myScenario$assignmentData = myScenario$assignmentData \%>\% filter(isFalseTrack == FALSE)
plot_scatterplot_with_density(myScenario, xValue = 'bearingError', yValue = 'downrangeError')
}
