% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_target_assignments_plotly.R
\name{plot_target_assignments_plotly}
\alias{plot_target_assignments_plotly}
\title{Plot target assignments}
\usage{
plot_target_assignments_plotly(scenario)
}
\arguments{
\item{scenario}{MUST contain targetTruth and assignmentData (from target_assignment())}
}
\value{
ggplot object
}
\description{
Figure that shows which sensor points were assigned to which targets.  Thin lines connect the points the sensor point to the interpolated true position at that time.  It also indicates which points were determined to be false tracks.  This figure is very useful to show the differences between assignment methods but can be difficult to read when there are many targets and tracks in one figure. The figure plots all of the targets as well as all of the tracks that were not determined to be false tracks. False track points are shown as black dots.
 Plotly figures are interactive, and users can zoom in, hover over data points for more info, and click to hide tracks or targets.
}
\examples{
plot_target_assignments_plotly(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100))
}
