% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track_status_plotly.R
\name{plot_track_status_plotly}
\alias{plot_track_status_plotly}
\title{Visualize when targets were tracked with plotly}
\usage{
plot_track_status_plotly(scenario)
}
\arguments{
\item{scenario}{MUST contain targetTruth and assignmentData (from target_assignment())}
}
\value{
plotly object
}
\description{
Plot truth data, color coded to show whether target was tracked.  Colors indicate whether targets were tracked with zero, one, or multiple tracks at a time.  Plotly figures are interactive, and users can zoom in, hover over data points for more info, and click to hide tracks or targets.
}
\examples{
plot_track_status_plotly(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100))
}
