% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tracks_per_target.R
\name{plot_tracks_per_target}
\alias{plot_tracks_per_target}
\title{Plot tracks per target}
\usage{
plot_tracks_per_target(scenario)
}
\arguments{
\item{scenario}{must contain assignmentData, the output of target_assignment()}
}
\value{
ggplot object
}
\description{
This figure shows the number of tracks per target at each time that the sensor gave a reading.
}
\examples{
plot_tracks_per_target(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100))
}
