% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_performance.R
\name{summarize_performance}
\alias{summarize_performance}
\title{Summarize sensor performance}
\usage{
summarize_performance(scenario)
}
\arguments{
\item{scenario}{must contain assignmentData, the output of target_assignment()}
}
\value{
data frame containing a summary of detection range and time, average error, average downrange error, average bearing error, and total number of sensor points for each target.  The data frame contains the collowing variables:

\itemize{
 \item{"Tgt Name": Name of target}
 \item{"Detection Range (km)": range at which the target was first detected}
 \item{"Detection Time": time at which the target was first detected}
 \item{"Avg Abs Position Error (m)": average absolute difference between target location and location of assigned sensor point}
 \item{"Avg Abs Downrange Error (m)": average absolute difference between target range and range of assigned sensor point}
 \item{"Avg Abs Bearing Error (deg)": average absolute difference between target bearing and bearing of assigned sensor point}
 \item{"# of sensor points": total number of sensor points assigned to the target}}
}
\description{
Table showing when targets were detected
}
\examples{
summarize_performance(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100))
}
