% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_time_tracked.R
\name{summarize_time_tracked}
\alias{summarize_time_tracked}
\title{Summarize amount of time each target was tracked}
\usage{
summarize_time_tracked(scenario)
}
\arguments{
\item{scenario}{must contain assignmentData, the output of target_assignment()}
}
\value{
Data frame containing a summary of how much time each target was tracked.  The data frame contains the collowing variables:

\itemize{
 \item{"tgt" - Target name/identifier}
 \item{"Total Time Tracked" - The total amount of time that the target had at least one valid track}
 \item{"# of Unique Tracks" - The number of segments that at one point were assigned to the target. If a track is assigned to target A, then jumps to target B, then returns to target A, then that would count as two unique tracks on target A.}
 \item{"# of Tracked Sections" - Total number of sections covering the track. If target A is covered by track 1, and then track 2 covers target A before track 1 ends then that will count as only one tracked section.}
 \item{"Time Detected" - the first time that a target is detected}
 \item{"Tracking Ended" - the last time that a target is detected}}
}
\description{
Table showing when targets were detected
}
\examples{
summarize_time_tracked(scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100))
}
