% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_assignment_gauss.R
\name{target_assignment.gauss}
\alias{target_assignment.gauss}
\title{Target assignment (Gaussian window method) - DO NOT DIRECTLY CALL THIS FUNCTION}
\usage{
target_assignment.gauss(
  targetTrackDistance,
  cutoff,
  windowSize,
  parallel = FALSE
)
}
\arguments{
\item{targetTrackDistance}{the output of target_track_distance()}

\item{cutoff}{numeric vector of length 1, 2, or 3 specifying the conditions under which a point is determined to be a false track. The different lengths imply different cutoff conditions:
\itemize{
 \item{One number (vector length 1): points are false tracks if the location error (difference between sensor point and closest target) is greater than this value in meters}
 \item{Two numbers (vector length 2): first number is cutoff in meters of the 2D error (sqrt(locationError^2 - altError^2)). Second number is cutoff in meters of the absolute value of the altitude error (sensor altitude reading - true target altitude)}
 \item{Three numbers (vector length 3): first number is range error in meters, second number is bearing error in degrees, third number is altitude error in meters}}}

\item{windowSize}{standard deviation of the weights to be used in the weighted mean (in same units as original sensorData input, likely POSIX).}

\item{parallel}{when TRUE, will run target assignment on multiple cores}
}
\value{
A data frame with the same number of observations as the sensor data that was originally passed to target_track_distance(). Each sensor point in the original data set has been assigned to a target using the method specified by the user, and the output data provides the following summary statistics:
\itemize{
 \item{lonError: difference in longitude between the sensor point and assigned target at the time of the sensor point}
 \item{latError: difference in latitude between the sensor point and assigned target at the time of the sensor point}
 \item{altError: difference in altitude between the sensor point and assigned target at the time of the sensor point}
 \item{bearingToTarget: bearing (azimuth) to target from ownship at the time of the sensor point}
 \item{trackNum: the track number associated with this sensor data point}
 \item{tgtAssigned: the ID of the target assigned to this sensor point}
 \item{locationError: distance between sensor point and target at the given time}
 \item{pointIndex: index referring to which of the original sensor data points this target-track pair refers. Necessary for comparisons in target_assignment()}
 \item{time: the time that this sensor point was recorded}
 \item{bearingError: difference in bearing between the sensor point and target at the time of the sensor point}
 \item{downrangeError: difference in range to ownship between the sensor point and target}
 \item{lon: longitude of the sensor point}
 \item{lat: latitude of the sensor point}
 \item{alt: altitude of the sensor point}
 \item{rangeToShip: range from target to ownship at the time of the sensor data point}
 \item{targetAspect: target aspect (as seen from ownship) at the time of the sensor data point}
 \item{meanLocationError: (only in square and gauss window methods) mean distance between sensor point and target for all of the points included in the window}
 \item{isFalseTrack: boolean indicating whether a point is outside the cutoff and therefore considered a false track}
 \item{tgtXtrack: factor expressing the truthID.trackNum interaction}
 \item{segmentNumber: an integer counting the number of times during which a single track is assigned to a particular target}
 }
}
\description{
This function IS NOT intended to be called by the user.  The user is supposed to call target_assignment(method = 'windowGauss') to access this function.  See \code{\link{target_assignment}} for additional details.

The Gaussian window method (method = "windowGauss") is closely related to the square window method. Instead of taking an average of points within a given window for purposes of assigning a sensor point to a target, the gauss window method takes a normally-weighted average, with the windowSize argument as the standard deviation in time of the distribution. For each point NAUTILUS:

1. Calculates the weight for all points based on their time, treating the time at the given point as the center of the Gaussian (i.e., if the given point in a track is at time 60 and the windowSize parameter of 10 gives weights for all of the points as determined by dnorm( allTimesInTrack , mean = 60, sd =10) )

2. The average distance between a track and a target is calculated as the weighted mean of the distance between all of the points in the track compared to the target at each time, weighted by the Gaussian function determined above

3. NAUTILUS assigns the track to the closest target or specifies that it is considered a false track if it is outside the user specified cutoff.  See \code{\link{get_isFalseTrack}} for additional details on the cutoff
}
