% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_assignment_secondpass.R
\name{target_assignment_secondpass}
\alias{target_assignment_secondpass}
\title{Second-pass method of target assignment}
\usage{
target_assignment_secondpass(scenario, method, excludeFirstPass = FALSE, ...)
}
\arguments{
\item{scenario}{scenario must contain the output of target_track_assignment() (the result of the first pass)}

\item{method}{string specifying which subfunction to call. Options include:
\itemize{
 \item{"point": point method of target assignment. Calls target_assignment.point().  Requires additional parameters "cutoff".  See \code{\link{target_assignment.point}} for additional details}
 \item{"wholeTrack": whole track method of target assignment. Calls target_assignment.track().  Requires additional parameters "cutoff".  See \code{\link{target_assignment.track}} for additional details}
 \item{"windowSquare": square window method of target assignment. Calls target_assignment.window(), requires additional parameters "cutoff" and "windowSize".  optional parameter "parallel" will parallelize computation across multiple cores.  default for this is FALSE.  See \code{\link{target_assignment.window}} for additional details}
 \item{"windowGauss": Gaussian window method of target assignment. Calls target_assignment.gauss().  Requires additional parameters "cutoff" and "windowSize".  optional parameter "parallel" will parallelize computation across multiple cores.  default for this is FALSE.  See \code{\link{target_assignment.gauss}} for additional details}
 \item{"user": user-specified method of target assignment. Calls target_assignment.user().  Requires additional parameters "userAssignedVector".  See \code{\link{target_assignment.user}} for additional details}
}}

\item{excludeFirstPass}{(Default = FALSE) determines whether to include the false tracks from the first pass in the final data frame.}

\item{...}{additional commands to be passed to each subfunction ( target.assignment.XXXXX() ). For example, all of the subfunctions require a cutoff parameter and the tho window methods (method = "windowSquare" or "windowGauss") each require the window size (windowSize) to be specified.  See \code{\link{target_assignment}} for addiitonal details.}
}
\value{
A data frame with the same number of observations as the sensor data that was originally passed to target_track_distance(). Each sensor point in the original data set has been assigned to a target using the method specified by the user, and the output data provides the following summary statistics:
\itemize{
 \item{lonError: difference in longitude between the sensor point and assigned target at the time of the sensor point}
 \item{latError: difference in latitude between the sensor point and assigned target at the time of the sensor point}
 \item{altError: difference in altitude between the sensor point and assigned target at the time of the sensor point}
 \item{bearingToTarget: bearing (azimuth) to target from ownship at the time of the sensor point}
 \item{trackNum: the track number associated with this sensor data point}
 \item{tgtAssigned: the ID of the target assigned to this sensor point}
 \item{locationError: distance between sensor point and target at the given time}
 \item{pointIndex: index referring to which of the original sensor data points this target-track pair refers. Necessary for comparisons in target_assignment()}
 \item{time: the time that this sensor point was recorded}
 \item{bearingError: difference in bearing between the sensor point and target at the time of the sensor point}
 \item{downrangeError: difference in range to ownship between the sensor point and target}
 \item{lon: longitude of the sensor point}
 \item{lat: latitude of the sensor point}
 \item{alt: altitude of the sensor point}
 \item{rangeToShip: range from target to ownship at the time of the sensor data point}
 \item{targetAspect: target aspect (as seen from ownship) at the time of the sensor data point}
 \item{meanLocationError: (only in square and gauss window methods) mean distance between sensor point and target for all of the points included in the window}
 \item{isFalseTrack: boolean indicating whether a point is outside the cutoff and therefore considered a false track}
 \item{tgtXtrack: factor expressing the truthID.trackNum interaction}
 \item{segmentNumber: an integer counting the number of times during which a single track is assigned to a particular target}
 }
}
\description{
There are several reasons why one may want to run target_assignment() twice.  For example, for large data sets with a large number of false tracks (clutter), running window methods may be very slow.  It may be possible to speed up computation by first running a fast method of target_assignment() ()point or whole track) with a large cutoff value, then returning to run the slower window method on the reduced data set.  As another example, a user may want to ensure that they keep all points within a track (as the method = 'wholeTrack' would do), however they also want a track to be able to switch targets (as the other methods can).  In this case, the user might want to run the whole track method first with a medium-sized cutoff to remove all of the false tracks, and then return and use one of the other methods with a large cutoff to ensure that none of the remaining points are thrown out while still allowing them to switch targets.

This function can be used to easily run \code{\link{target_assignment}} for a second pass after it has already been run once.  The first pass should be designed to remove as much of the false tracks as possible without interfering with data that the user would want to keep.  It will then run only those points NOT determined to be false tracks through the second pass.  The points eliminated in the first pass will be included in the final data frame so that no data points are lost during this process (unless the user specifies otherwise, perhaps to reduce the overall size of the data for plotting).

The first pass should properly be a faster method to properly take advantage of the two-pass technique, therefore we recommend using method='point' or method='wholeTrack' for the first pass before using this function.
}
\examples{


# Example 1: speeding up the gaussian window method by
# eliminating some false tracks in the first pass

    target_assignment(scenario=scenarioMaker::example2_scenario,
     method = 'point', cutoff = 300) \%>\%
    target_assignment_secondpass(method = 'windowGauss',
     cutoff = 100, windowSize = 30)


# Example 2: Removing all whole tracks that are false tracks
# and keeping all of the data points for tracks that are not
# false tracks.  In the second pass, a large cutoff prevents
# any further points from being removed while allowing for
# tracks to switch targets

    target_assignment(scenario=scenarioMaker::example2_scenario,
     method = 'wholeTrack', cutoff = 200) \%>\%
    target_assignment_secondpass(method = 'windowGauss',
     cutoff = 5000, windowSize = 30)
}
