% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateData.R
\name{annotateData}
\alias{annotateData}
\title{Annotate Dataset with Events or Behaviors of Interest}
\usage{
annotateData(
  data,
  event.data,
  column.name = "event",
  datetime.col = "datetime"
)
}
\arguments{
\item{data}{A data frame containing the original dataset with a \code{datetime} column.}

\item{event.data}{A data frame containing the start and end times of events/behaviors (in \code{POSIXct} format).}

\item{column.name}{A character string specifying the name for the new binary column.}

\item{datetime.col}{A character string specifying the name of the column in \code{data} containing the datetime values.}
}
\value{
A data frame with an additional binary column indicating whether each row falls within the specified event/behavior.
}
\description{
This function takes an original dataset and a data frame with event start and end times, then creates a new binary column
in the dataset, which marks rows that fall within the specified event or behavior periods as \code{1}, and others as \code{0}.
}
