% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotateData.R
\name{annotateData}
\alias{annotateData}
\title{Annotate Dataset with Events or Behaviors of Interest}
\usage{
annotateData(
  data,
  annotations,
  id.col = "ID",
  datetime.col = "datetime",
  event.col = "event",
  start.col = "start",
  end.col = "end",
  selected.events = NULL
)
}
\arguments{
\item{data}{A list of data tables/data frames, one for each individual, or a single aggregated data table/data frame containing data from multiple animals.}

\item{annotations}{A data frame containing the start and end times of events/behaviors (columns specified by \code{start.col} and \code{end.col})
and an individual ID column (specified by \code{id.col}).}

\item{id.col}{A character string specifying the name of the column in both \code{data} and \code{annotations} containing animal IDs.}

\item{datetime.col}{A character string specifying the name of the column in \code{data} containing the datetime values.}

\item{event.col}{A character string specifying}

\item{start.col}{A character string specifying the name of the column in \code{annotations} containing event start times.}

\item{end.col}{A character string specifying the name of the column in \code{annotations} containing event end times.}

\item{selected.events}{A vector of event types to be annotated. If NULL (default), all events will be annotated.}
}
\value{
A list of data frames (one for each individual) with binary columns indicating the presence (1) or absence (0) of each specified event.
}
\description{
This function annotates a list of individual data frames (or a single aggregated data frame) with event or behavior information.
It links time-bound events (e.g., feeding, social interactions) to sensor data collected from individual animals,
with the event annotations typically derived from a detailed review of footage recorded by camera tags.
The function scans the dataset(s) for each animal, checking whether each row's timestamp falls within the defined time intervals of
specific events listed in the provided annotations data frame, marking rows that fall within the specified event or behavior
periods as \code{1} and those outside the periods as \code{0}.
}
