% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFeatures.R
\name{extractFeatures}
\alias{extractFeatures}
\title{Extract Features from a Sliding Window}
\usage{
extractFeatures(
  data,
  variables = c("depth", "surge", "pitch"),
  metrics = c("mean", "sd"),
  window.size = 5,
  aggregate = FALSE,
  datetime.col = "datetime",
  response.col = NULL
)
}
\arguments{
\item{data}{A data frame containing the input data. Must include the variables specified in \code{variables}.}

\item{variables}{A character vector of column names from \code{data} for which to calculate metrics.}

\item{metrics}{A character vector specifying the metrics to calculate. Supported metrics: "mean", "sd", "min", "max", "sum".}

\item{window.size}{An integer specifying the size of the sliding or aggregation window.}

\item{aggregate}{Logical. If \code{TRUE}, aggregates data into distinct, non-overlapping windows. If \code{FALSE}, retains temporal resolution.}

\item{datetime.col}{A character string specifying the column in \code{data} containing datetime information.}

\item{response.col}{(Optional) A character string specifying the column in \code{data} containing response or annotated labels (e.g., feeding events).
It must be binary (i.e., containing values 0 or 1).}
}
\value{
A data frame containing the calculated features. If \code{aggregate = TRUE}, the output will have fewer rows,
with one row per window. If \code{aggregate = FALSE}, the output retains the original number of rows, with
calculated metrics for each window. If a \code{datetime} column exists in the input data, it will be adjusted to match
the window aggregation, and the corresponding timestamp will be retained or adjusted based on the window size.
The response column, if specified, will also be aggregated or retained according to the same rules.
}
\description{
This function calculates specified metrics (e.g., mean, standard deviation) for selected variables over a sliding window.
It can either retain the temporal resolution of the dataset or aggregate data into distinct, non-overlapping windows.
This is useful for preparing the dataset for machine learning or other analytical methods that require a structured set of features.
}
