% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDeploymentData.R
\name{filterDeploymentData}
\alias{filterDeploymentData}
\title{Filter deployment periods from a given dataset}
\usage{
filterDeploymentData(
  data,
  first.datetime,
  last.datetime,
  psat.data = NULL,
  depth.threshold = 3.5,
  variance.threshold = 6,
  max.changepoints = 6,
  id.col = "ID",
  datetime.col = "datetime",
  depth.col = "depth",
  plot.vars = c("ax", "pitch"),
  plot = TRUE
)
}
\arguments{
\item{data}{A data frame containing the dataset, including a column for 'depth' and 'datetime'.}

\item{first.datetime}{A datetime object indicating the start of the first deployment period.}

\item{last.datetime}{A datetime object indicating the end of the last deployment period.}

\item{psat.data}{An optional data frame containing PSAT data for additional plotting (default is NULL).}

\item{depth.threshold}{A numeric value for the threshold depth to distinguish deployment periods (default is 3.5).}

\item{variance.threshold}{A numeric value for the threshold variance to distinguish deployment periods (default is 6).}

\item{max.changepoints}{An integer specifying the maximum number of changepoints to detect (default is 6).}

\item{id.col}{A string representing the column name for the ID field (default is "ID").}

\item{datetime.col}{A string representing the column name for the datetime field (default is "datetime").}

\item{depth.col}{A string representing the column name for the depth field (default is "depth").}

\item{plot.vars}{A character vector specifying the variables to plot, such as acceleration ("ax") or pitch (default is c("ax", "pitch")).}

\item{plot}{A boolean value indicating whether to generate the plot (default is TRUE).}
}
\value{
A filtered data frame containing the deployment period between the detected attachment and popup times.
}
\description{
This function identifies deployment periods by analyzing changes in depth data using binary segmentation
to detect change points in both mean and variance. It then extracts the periods between the pre-deployment
and post-deployment phases based on the specified datetime range.
}
\details{
The function uses a binary segmentation method (\code{cpt.meanvar}) to detect changes in depth and variance,
and it allows users to specify the maximum number of changepoints detected. The function also provides options
to visualize the data with custom plot options.
}
