% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterDeploymentData.R
\name{filterDeploymentData}
\alias{filterDeploymentData}
\title{Filter deployment periods from a given dataset}
\usage{
filterDeploymentData(
  data,
  depth.threshold = 3.5,
  variance.threshold = 6,
  max.changepoints = 6,
  id.col = "ID",
  datetime.col = "datetime",
  depth.col = "depth",
  plot.metrics = c("pitch", "sway"),
  plot.metrics.labels = NULL,
  display.plots = FALSE,
  save.plots = TRUE
)
}
\arguments{
\item{data}{A list of data tables/data frames, one for each individual, or a single aggregated data table/data frame
containing data from multiple animals. The output of the \link{processTagData} function is recommended, as it formats
the data appropriately for further analysis.}

\item{depth.threshold}{A numeric value specifying the minimum mean depth to classify a segment as part of
the deployment period (default is 3.5).}

\item{variance.threshold}{A numeric value specifying the minimum variance in depth measurements to classify
a segment as part of the deployment period. Segments with variance below this value are considered spurious (default is 6).}

\item{max.changepoints}{An integer specifying the maximum number of changepoints to detect (default is 6).
This parameter is passed to the \code{\link[changepoint]{cpt.meanvar}} function.}

\item{id.col}{A string representing the column name for the ID field (default is "ID").}

\item{datetime.col}{A string specifying the name of the column that contains timestamps for each observation.
This column must be in "POSIXct" format for proper processing (default is "datetime").}

\item{depth.col}{A string specifying the name of the column that contains depth measurements. Depth data is
used for detecting deployment periods (default is "depth").}

\item{plot.metrics}{A character vector of column names indicating additional metrics (e.g., acceleration or pitch)
to include in the visualization. These metrics are plotted alongside the depth data to aid in visually reviewing
the deployment period assignments (default is \code{c("pitch", "sway")}).}

\item{plot.metrics.labels}{A character vector specifying custom titles for the metrics plotted in the output.
If set to NULL, the column names provided in \code{plot.metrics} will be used as labels (default is NULL).}

\item{display.plots}{A logical value indicating whether the diagnostic plots should be displayed.
These plots generate diagnostic visuals showing depth and additional metrics,
assisting users in reviewing the extracted deployment periods.
If set to \code{TRUE}, the plots will be shown in the active graphics device.
Note: If set to \code{TRUE}, the code can take longer to run due to the delay in plotting extensive data series to the screen.
If set to \code{FALSE}, no plots will be displayed, even if they are saved.
Default is \code{FALSE}.}

\item{save.plots}{A logical value indicating whether the diagnostic plots should be saved.
If set to \code{TRUE}, the plots will be recorded and stored in the \code{diagnostic_plots} list.
The saved plots can later be displayed using \code{replayPlot()}.
These plots generate diagnostic visuals showing depth and additional metrics, assisting users in reviewing the extracted deployment periods.
If set to \code{FALSE}, no plots will be saved.
Default is \code{FALSE}.}
}
\value{
A filtered data frame containing the deployment period between the detected attachment and popup times.
}
\description{
This function identifies deployment periods by analyzing changes in depth data using binary segmentation
to detect change points in both mean and variance. It then extracts the periods between the pre-deployment
and post-deployment phases based on the specified datetime range.
}
\details{
The function uses a binary segmentation method (\code{cpt.meanvar}) to detect changes in depth and variance,
and it allows users to specify the maximum number of changepoints detected.

If the sampling frequency is greater than 1 Hz, the function downsamples the data to 1 Hz by rounding
datetime values to the nearest second and computing the mean for each second. This is done to improve speed
and avoid memory bottlenecks when handling large datasets.

The function also provides options to visualize the data with custom plot options.
}
\seealso{
\link{processTagData}, \code{\link[changepoint]{cpt.meanvar}}.
}
