% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchVideo.R
\name{launchVideo}
\alias{launchVideo}
\title{This function searches for the video file that contains the given datetime
within its start and end times. If a matching video is found, it opens
the video using VLC at the time corresponding to the given datetime.}
\usage{
launchVideo(datetime, video.metadata, vlc.path = NULL)
}
\arguments{
\item{datetime}{A \code{POSIXct} object representing the datetime to find
within the video metadata.}

\item{video.metadata}{A \code{data.frame} containing the metadata for each
video, as returned by \code{getVideoMetadata}.}

\item{vlc.path}{(Optional) The full path to the VLC application. If not
provided, a default path will be used based on the user's operating system.}
}
\value{
The function opens the corresponding video at the specified datetime
using VLC. If no video matches the datetime, an appropriate message is printed.
}
\description{
This function searches for the video file that contains the given datetime
within its start and end times. If a matching video is found, it opens
the video using VLC at the time corresponding to the given datetime.
}
