% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compare_track_status.R
\name{plot_compare_track_status}
\alias{plot_compare_track_status}
\title{Compares the track status from different assignmentDatas}
\usage{
plot_compare_track_status(assignmentDataList, nameList, truthData,
  showFalseTracks = FALSE)
}
\arguments{
\item{assignmentDataList}{list of two assignmentData dataframes (the output of target_assignment())}

\item{nameList}{list of two names for plotting (in the same order as assignmentDataList)}

\item{truthData}{truthData used in creating assignmentData sets}

\item{showFalseTracks}{will plot false tracks for comparison (default=FALSE)}
}
\value{
ggplot object
}
\description{
This figure compares the target assignments from two methods.  Each subfigure is similar to plot_track_status().  Areas of disagreement between the two methods are shown in bold.  If there are differences in points flagged as false tracks they are shown as triangles.
}
\examples{
pointScenario=scenarioMaker::example1_scenario \%>\% target_assignment("point",cutoff=100)
gaussScenario=scenarioMaker::example1_scenario \%>\% target_assignment("windowGauss",cutoff=100, window=10)
plot_compare_track_status(list(pointScenario$assignmentData, gaussScenario$assignmentData),list("Point Method", "Gauss Window Method"), scenarioMaker::example1_scenario$targetTruth)
}
