% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModelDiagnostics.R
\name{plotModelDiagnostics}
\alias{plotModelDiagnostics}
\title{Plot Model Diagnostics}
\usage{
plotModelDiagnostics(
  model,
  test.data,
  title.cex = 1,
  lab.cex = 0.9,
  axis.cex = 0.8
)
}
\arguments{
\item{model}{A trained random forest model object. This model is expected to be trained on a dataset where the dependent variable is a factor.}

\item{test.data}{A data frame representing the testing dataset used to evaluate the model. Should include the dependent variable as the first column.}

\item{title.cex}{A numeric value indicating the text size for plot titles. Default is 1.}

\item{lab.cex}{A numeric value indicating the text size for axis labels. Default is 0.9.}

\item{axis.cex}{A numeric value indicating the text size for axis tick marks and bar labels. Default is 0.8.}
}
\description{
Generates a series of diagnostic plots for a machine learning model, including feature importance, ROC curve,
confusion matrix heatmap, and performance metrics.
}
\details{
The function produces the following four plots:
\itemize{
\item \strong{Feature Importance:} Displays a horizontal barplot of the feature importance values based on Mean Decrease Gini.
\item \strong{ROC Curve:} Plots the Receiver Operating Characteristic (ROC) curve with the Area Under Curve (AUC) displayed in the title.
\item \strong{Confusion Matrix Heatmap:} A heatmap representation of the confusion matrix, showing proportions for predicted vs. actual classifications.
\item \strong{Model Performance Metrics:} A barplot summarizing precision, recall, and F1 score calculated from the confusion matrix.
}
The function assumes the response variable in \code{test.data} is a factor.
}
