% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reencodeVideos.R
\name{reencodeVideos}
\alias{reencodeVideos}
\title{Reencode MOV Videos to HEVC Format}
\usage{
reencodeVideos(mov.directory, output.directory = mov.directory)
}
\arguments{
\item{mov.directory}{A string specifying the path to the directory containing \code{.mov} files to re-encode.}

\item{output.directory}{A string specifying the path to the directory where re-encoded \code{.mp4} files will be saved. Defaults to the same as \code{mov_directory}.}
}
\value{
Prints progress and time statistics to the console. The re-encoded videos are saved in the specified output directory.
}
\description{
This function re-encodes all \code{.mov} video files in a specified input directory to HEVC (H.265) format
using hardware acceleration (\code{hevc_videotoolbox}). The re-encoded videos are saved
as \code{.mp4} files in the specified output directory or the input directory by default.
}
\details{
\itemize{
\item The re-encoding is done using the \code{ffmpeg} tool with the following options:
\itemize{
\item Video codec: \code{hevc_videotoolbox} for hardware-accelerated encoding
\item Constant rate factor (\code{crf}) of 28 for a balance of quality and compression
\item Preset: \code{fast} for faster encoding speed
\item No audio (\code{-an}).
}
}
}
