% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTagData.R
\name{summarizeTagData}
\alias{summarizeTagData}
\title{Generate Summary Statistics for Multiple Individuals}
\usage{
summarizeTagData(data, id.col = "ID", datetime.col = "datetime")
}
\arguments{
\item{data}{A list of data tables/data frames, one for each individual, or a single aggregated data table/data frame containing data from multiple animals.}

\item{id.col}{Column name for the animal ID (default is "ID").}

\item{datetime.col}{Column name for the datetime (default is "datetime").}
}
\value{
A data frame with summary statistics for each animal, including the following columns:
\itemize{
\item \strong{ID}: The unique identifier for each individual animal.
\item \strong{Total duration (h)}: The total dataset duration, in hours (including pre- and post-deployment periods).
\item \strong{Deploy start}: The estimated start (attachment) datetime of the deployment.
\item \strong{Deploy end}: The estimated end (pop-up) datetime of the deployment.
\item \strong{Deploy duration (h)}: The duration of the deployment, in hours, calculated from the deployment start to deployment end.
\item \strong{Sampling freq (Hz)}: The sampling frequency of the original dataset in Hertz (Hz)..
\item \strong{Magnetic declination}: The magnetic declination value used to correct heading estimates.
\item \strong{Max Depth (m)}: The maximum depth recorded during the deployment, in meters.
\item \strong{Temp range (ºC)}: The temperature range (minimum and maximum temperatures recorded) during the deployment, in degrees Celsius (°C).
\item \strong{Fastloc GPS}: The number of "Fast GPS" positions recorded during the deployment (if any).
\item \strong{User Locs}: The number of user-defined GPS positions recorded during the deployment (if any).
}
}
\description{
This function takes processed data from multiple individuals (either a list with a data.table/data.frame for each animal or an aggregated data.table/data.frame containing data from multiple animals)
and outputs a summary table with key statistics for each individual.
}
\seealso{
\link{processTagData}, \link{filterDeploymentData}.
}
