% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nautilus-class.R
\name{summary.nautilus_tag}
\alias{summary.nautilus_tag}
\title{Summarise a nautilus_tag}
\usage{
\method{summary}{nautilus_tag}(object, ...)
}
\arguments{
\item{object}{A \code{nautilus_tag} object.}

\item{...}{Unused.}
}
\value{
A one-row data.frame of per-deployment summary statistics.
}
\description{
Summarise a nautilus_tag
}
